Developer

See merge request mis/mis-ui!62

+++ src/views/rental/RentalManage.js
@@ -9,6 +9,7 @@ import { AuthContext } from '../../components/AuthContext';
 import '../../styles/common/Page.css';
 import '../../styles/rental/RentalManage.css';
 import useRentalChange from '../../hooks/useRentalChange';
+import Loading from '../../components/common/Loading';
 
 
 
@@ -20,6 +21,7 @@ function RentalManage() {
   const [isBulkUpdateModalVisible, setIsBulkUpdateModalVisible] = useState(false);
   const [selectedRental, setSelectedRental] = useState(null); 
   const [rentalDetails, setRentalDetails] = useState([]);  
+  const [loading, setLoading] = useState(false);
 
   const dragStartIndex = useRef(null);
   const dragEndIndex = useRef(null);
@@ -37,6 +39,7 @@ function RentalManage() {
   };
 
   const fetchRentalData = useCallback(async () => {
+    setLoading(true)
     try {
       const response = await axios.get(`/api/rentalList/center`, {
         params: { instCd: auth.instCd },
@@ -52,6 +55,8 @@ function RentalManage() {
       setSelectedRows([]);
     } catch (error) {
       console.error('센터 렌탈현황을 불러오는데 실패했습니다.', error);
+    } finally {
+      setLoading(false)
     }
   }, [auth.instCd, setSelectedRows]);
 
@@ -128,13 +133,23 @@ function RentalManage() {
         alert("삭제할 항목을 선택하세요.");
         return;
     }
-  
+
+    const completedItems = selectedRows.filter(rowId => {
+        const selectedItem = rentalDetails.find(item => item.detailId === rowId);
+        return selectedItem && selectedItem.status === '완료';
+    });
+
+    if (completedItems.length > 0) {
+        alert("완료된 내역은 삭제할 수 없습니다.");
+        return;
+    }
+
     try {
         for (const detailId of selectedRows) {
             await axios.delete(`/api/rental/`, { params: { detailId } });
         }
         alert('선택된 항목이 삭제되었습니다.');
-  
+
         setRentalDetails(prevDetails => {
             const updatedDetails = prevDetails
                 .filter(item => !selectedRows.includes(item.detailId))
@@ -145,13 +160,13 @@ function RentalManage() {
             
             return updatedDetails;
         });
-  
+
         setSelectedRows([]); 
     } catch (error) {
         console.error('렌탈현황 정보를 삭제하는 중 에러 발생:', error);
         alert('삭제에 실패했습니다.');
     }
-  };
+};
 
   const handleFinishButtonClick = async () => {
     if (selectedRows.length === 0) {
@@ -258,8 +273,8 @@ function RentalManage() {
     },
     { header: 'NO', accessor: 'no' },
     { header: '상태', accessor: 'status' },
-    { header: '제품군', accessor: 'category' },
     { header: '업체명', accessor: 'companyNm' },
+    { header: '제품군', accessor: 'category' },
     { header: '계약번호', accessor: 'contractNum' },
     { header: '모델명', accessor: 'modelNm' },
     { header: '설치일자', accessor: 'installDate' },
@@ -276,6 +291,10 @@ function RentalManage() {
         <div className="rental-content-inner">
           <h2>렌탈현황 관리표</h2>
           <Breadcrumb items={['자산 관리', '렌탈현황 관리표']} />
+          {loading ? (
+          <Loading />
+        ) : (
+          <>
           <div className="rental-tables-section">
             <div className="rental-details-content">
               <div className="rental-header-buttons">
@@ -300,6 +319,8 @@ function RentalManage() {
               </div>
             </div>
           </div>
+          </>
+        )}
         </div>
       </div>
       <RentalAddModal
